output with cgroup id -> u32

+++ eBPF.py
@@ -15,7 +15,7 @@ BPF_PERF_OUTPUT(events);
 int sgettimeofday(struct pt_regs *ctx) {
     struct data_t data = {};
     u64 id = bpf_get_current_pid_tgid();
-    u64 cgroup_id = bpf_get_current_cgroup_id();
+    u32 cgroup_id = bpf_get_current_cgroup_id();
     data.cgroup = cgroup_id;
     data.pid = id >> 32;
     data.syscallnumber = 0;
